+++ src/EntityGraphQL/Compiler/GqlNodes/GraphQLQueryStatement.cs
@@ -1,8 +1,6 @@
 using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq.Expressions;
-using System.Threading;
 using System.Threading.Tasks;
 using EntityGraphQL.Directives;
 using EntityGraphQL.Schema;
@@ -14,27 +12,27 @@ public class GraphQLQueryStatement : ExecutableGraphQLStatement
     public GraphQLQueryStatement(ISchemaProvider schema, string? name, Expression nodeExpression, ParameterExpression rootParameter, Dictionary<string, ArgType> variables)
         : base(schema, name, nodeExpression, rootParameter, variables) { }
 
-    public override Task<ConcurrentDictionary<string, object?>> ExecuteAsync<TContext>(
-        TContext? context,
+    protected override ExecutableDirectiveLocation ExecutableDirectiveLocation => ExecutableDirectiveLocation.QUERY;
+    protected override ISchemaType SchemaType => Schema.GetSchemaType(Schema.QueryContextType, false, null)!;
+
+    protected override async Task<(object? data, bool didExecute, List<GraphQLError> errors)> ExecuteOperationField<TContext>(
+        CompileContext compileContext,
+        BaseGraphQLField field,
+        TContext context,
         IServiceProvider? serviceProvider,
         IReadOnlyDictionary<string, GraphQLFragmentStatement> fragments,
-        Func<string, string> fieldNamer,
         ExecutionOptions options,
-        QueryVariables? variables,
-        QueryRequestContext requestContext,
-        CancellationToken cancellationToken = default
+        IArgumentsTracker? docVariables
     )
-        where TContext : default
     {
-        Schema.CheckTypeAccess(Schema.GetSchemaType(Schema.QueryContextType, false, null), requestContext);
-
-        var result = new ConcurrentDictionary<string, object?>();
-        // pass to directives
-        foreach (var directive in Directives)
+        // apply directives
+        foreach (var directive in field.Directives)
         {
-            if (directive.VisitNode(ExecutableDirectiveLocation.QUERY, Schema, this, Arguments, null, null) == null)
-                return Task.FromResult(result);
+            if (directive.VisitNode(ExecutableDirectiveLocation.FIELD, Schema, field, Arguments, null, null) == null)
+                return (null, false, []);
         }
-        return base.ExecuteAsync(context, serviceProvider, fragments, fieldNamer, options, variables, requestContext, cancellationToken);
+        (var data, var didExecute) = await CompileAndExecuteNodeAsync(compileContext, context!, serviceProvider, fragments, field, docVariables);
+
+        return (data, didExecute, new List<GraphQLError>());
     }
 }