+++ src/EntityGraphQL/EntityGraphQLFieldException.cs
@@ -5,10 +5,12 @@ namespace EntityGraphQL;
 internal sealed class EntityGraphQLFieldException : Exception
 {
     public readonly string FieldName;
+    public readonly string[] Path;
 
-    public EntityGraphQLFieldException(string fieldName, Exception innerException)
+    public EntityGraphQLFieldException(string fieldName, string[]? path, Exception innerException)
         : base(null, innerException)
     {
         FieldName = fieldName;
+        Path = path ?? [];
     }
 }